[UPDATE] Work on campaigns

+++ src/routes/campaigns.lazy.tsx
@@ -2,6 +2,8 @@ import * as React from "react";
 import { createLazyFileRoute } from "@tanstack/react-router";
 import Session, { SessionAuth } from "supertokens-auth-react/recipe/session";
 import { ConfirmDialog, DialogOptions } from "primereact/confirmdialog";
+import { Dialog } from "primereact/dialog";
+import { Button } from "primereact/button";
 import { IoWarningSharp } from "react-icons/io5";
 import { toast } from "react-toastify";
 
@@ -16,26 +18,29 @@ export const Route = createLazyFileRoute("/campaigns")({
 function Campaigns() {
     const [campaigns, SetCampaigns] = React.useState<CampaignOverview[]>([]);
     const [activeCampaign, SetActiveCampaign] = React.useState<Campaign | null>(null);
-    const [dialogOptions, SetDialogOptions] = React.useState<DialogOptions>({ visible: false, label: "", message: "", onHide: () => { }, accept: () => { }, reject: () => { } });
+    const [confirmDialogOptions, SetConfirmDialogOptions] = React.useState<DialogOptions>({ visible: false, label: "", message: "", onHide: () => { }, accept: () => { }, reject: () => { } });
+    const [openCreationDialog, SetOpenCreationDialog] = React.useState<boolean>(false);
     const getCampaignsRef = React.useRef<number>(0);
+    const [LocalStringState1, SetLocalStringState1] = React.useState<string>("");
+    const [LocalStringState2, SetLocalStringState2] = React.useState<string>("");
 
     const getCampaign = (name: string) => {
         api.getCampaign(name).then((res) => {
             if (!res.Campaign) return;
             SetActiveCampaign(res.Campaign);
         });
-    }
+    };
 
     const selectCampaign = (campaignName: string) => {
         getCampaign(campaignName);
     };
 
     const deleteCampaign = (campaignName: string) => {
-        SetDialogOptions({
+        SetConfirmDialogOptions({
             visible: true,
             label: "Delete Campaign",
             message: `Are you sure you want to delete the campaign "${campaignName}"?`,
-            onHide: () => { SetDialogOptions({ ...dialogOptions, visible: false }) },
+            onHide: () => { SetConfirmDialogOptions({ ...confirmDialogOptions, visible: false }) },
             accept: () => {
                 api.deleteCampaign(campaignName).then((res: boolean) => {
                     if (res) {
@@ -53,6 +58,12 @@ function Campaigns() {
         });
     };
 
+    const createNewCampaign = () => {
+        SetLocalStringState1("");
+        SetLocalStringState2("");
+        SetOpenCreationDialog(true);
+    }
+
     React.useEffect(() => {
         if (getCampaignsRef.current === 0) {
             getCampaignsRef.current = 1;
@@ -60,7 +71,9 @@ function Campaigns() {
         }
     }, [getCampaignsRef]);
 
-    const PageTemplate = ({ children }: { children: React.ReactNode }) => {
+
+    // Campaign overview screen
+    if (!activeCampaign) {
         return (
             <SessionAuth
                 onSessionExpired={async () => {
@@ -68,18 +81,24 @@ function Campaigns() {
                     window.location.href = "/auth"
                 }}
             >
-                {children}
+                <h1>Campaigns</h1>
+                <div className="twelve columns">
+                    <h3 className="eight columns offset-by-one column">My Campaigns</h3>
+                    <button className="two columns" onClick={createNewCampaign}>Create New Campaign</button>
+                </div>
+                <div className="break" />
+                <CampaignsTable className="ten columns offset-by-one columns" campaigns={campaigns} nameCallback={selectCampaign} deleteCallback={deleteCampaign} />
                 <ConfirmDialog
-                    visible={dialogOptions.visible}
-                    onHide={() => { SetDialogOptions({ ...dialogOptions, visible: false }), dialogOptions.onHide() }}
-                    header={dialogOptions.label}
-                    message={dialogOptions.message}
+                    visible={confirmDialogOptions.visible}
+                    onHide={() => { SetConfirmDialogOptions({ ...confirmDialogOptions, visible: false }), confirmDialogOptions.onHide() }}
+                    header={confirmDialogOptions.label}
+                    message={confirmDialogOptions.message}
                     className="dialog"
                     focusOnShow={true}
-                    accept={dialogOptions.accept}
-                    reject={dialogOptions.reject}
-                    defaultFocus={dialogOptions.defaultFocus}
-                    icon={dialogOptions.icon}
+                    accept={confirmDialogOptions.accept}
+                    reject={confirmDialogOptions.reject}
+                    defaultFocus={confirmDialogOptions.defaultFocus}
+                    icon={confirmDialogOptions.icon}
                     modal={true}
                     acceptClassName="dialog-accept"
                     rejectClassName="dialog-reject"
@@ -87,25 +106,75 @@ function Campaigns() {
                     headerClassName="dialog-header"
                     contentClassName="dialog-content"
                 />
+                <Dialog
+                    visible={openCreationDialog}
+                    onHide={() => SetOpenCreationDialog(false)}
+                    header={"Create New Campaign"}
+                    className="dialog"
+                    focusOnShow={true}
+                    modal={true}
+                    maskClassName="dialog-mask"
+                    footer={<div>
+                        <Button label="Cancel" onClick={() => SetOpenCreationDialog(false)} />
+                        <Button label="Done" onClick={() => {
+                            api.createCampaign(new Campaign(LocalStringState1, LocalStringState2)).then((res) => {
+                                console.log(res);
+                                if (res) {
+                                    window.location.reload();
+                                    toast.success("Campaign created successfully.");
+                                }
+                                else {
+                                    toast.error("Failed to create campaign.");
+                                }
+                            }).catch((err) => {
+                                console.error(err);
+                                toast.error("Failed to create campaign. Please try again later.");
+                            });
+                            SetOpenCreationDialog(false);
+                        }} autoFocus />
+                    </div>}
+                >
+                    <div>
+                        <label htmlFor="campaign-name">Name</label>
+                        <input id="campaign-name" type="text" value={LocalStringState1} onChange={(e) => SetLocalStringState1(e.target.value)} placeholder="Campaign Name" />
+                        <br />
+                        <label htmlFor="campaign-description">Description</label>
+                        <input id="campaign-description" type="text" value={LocalStringState2} onChange={(e) => SetLocalStringState2(e.target.value)} placeholder="Campaign Description" />
+                        <br />
+                    </div>
+                </Dialog>
             </SessionAuth>
-        );
-    }
-
-    // Campaign overview screen
-    if (!activeCampaign) {
-        return (
-            <PageTemplate>
-                <h1>Campaigns</h1>
-                <CampaignsTable className="eight columns offset-by-two columns" campaigns={campaigns} nameCallback={selectCampaign} deleteCallback={deleteCampaign} />
-            </PageTemplate>
         );
     }
 
     // Campaign details screen
     return (
-        <PageTemplate>
+        <SessionAuth
+            onSessionExpired={async () => {
+                await Session.signOut()
+                window.location.href = "/auth"
+            }}
+        >
             <p>{activeCampaign.Name}</p>
             {/* TODO */}
-        </PageTemplate>
+            <ConfirmDialog
+                visible={confirmDialogOptions.visible}
+                onHide={() => { SetConfirmDialogOptions({ ...confirmDialogOptions, visible: false }), confirmDialogOptions.onHide() }}
+                header={confirmDialogOptions.label}
+                message={confirmDialogOptions.message}
+                className="dialog"
+                focusOnShow={true}
+                accept={confirmDialogOptions.accept}
+                reject={confirmDialogOptions.reject}
+                defaultFocus={confirmDialogOptions.defaultFocus}
+                icon={confirmDialogOptions.icon}
+                modal={true}
+                acceptClassName="dialog-accept"
+                rejectClassName="dialog-reject"
+                maskClassName="dialog-mask"
+                headerClassName="dialog-header"
+                contentClassName="dialog-content"
+            />
+        </SessionAuth>
     );
 }
